remove fallback date for `nextRunAt`
This just creates wrong data. Old jobs that have been executed already and will not be executed again appear as jobs that will run at the current date (as their date is null). It should just be null instead like it is written in the persistence layer. No need for a fallback.